complete migration of emacs from brew to nix

+++ darwin-configuration.nix
@@ -36,6 +36,7 @@
     "coreutils"
     "editorconfig"
     "gcc"
+    "libtool"
     "pngpaste"
     "podman"
     "podman-compose"
@@ -50,7 +51,6 @@
     "font-jetbrains-mono-nerd-font"
     "hazeover"
     "karabiner-elements"
-    "libtool"
     "logitech-options"
     "lulu"
     "numi"
@@ -61,17 +61,13 @@
     "zsa-wally"
   ];
   homebrew.taps = [
-    "homebrew/cask-versions"
     "homebrew/cask-drivers"
     "homebrew/cask-fonts"
+    "homebrew/cask-versions"
     "homebrew/services"
-    # "d12frosted/emacs-plus"
   ];
   homebrew.onActivation.cleanup = "zap";
   homebrew.onActivation.upgrade = true;
-  # homebrew.extraConfig = ''
-  #   brew "emacs-plus@29", args: ["with-dbus", "with-mailutils", "with-no-frame-refocus", "with-xwidgets", "with-imagemagick", "with-native-comp", "with-nobu417-big-sur-icon", "with-poll"]
-  # '';
   homebrew.masApps = {
     "Amphetamine" = 937984704;
     "Bitwarden" = 1352778147;
@@ -85,8 +81,8 @@
     "Noir" = 1592917505;
     "Sorted" = 1306893526;
     "SponsorBlock for YouTube - Skip Sponsorships" = 1573461917;
-    "uBlacklist for Safari" = 1547912640;
     "Wipr" = 1320666476;
+    "uBlacklist for Safari" = 1547912640;
   };
 
   fonts.fontDir.enable = true;
@@ -125,14 +121,17 @@
     rm -f /etc/shells
   '';
   system.activationScripts.postActivation.text = ''
-  #   if [ ! -d /Applications/Sorted.app ]; then
-  #       mv /Applications/Sorted* /Applications/Sorted.app
-  #   fi
+    #   if [ ! -d /Applications/Sorted.app ]; then
+    #       mv /Applications/Sorted* /Applications/Sorted.app
+    #   fi
     # Find the latest Emacs.app in the Nix store
     EMACS_APP=$(ls -dt /nix/store/*emacs*/Applications/Emacs.app | head -n 1)
 
+    # Change icon
+    cp /Users/sebastian/.nixpkgs/custom/Emacs.icns $EMACS_APP/Contents/Resources/Emacs.icns
+
     # Create a symlink to it in /Applications
-    unlink /Applications/Emacs.app
+    test -f /Applications/Emacs.app && unlink /Applications/Emacs.app
     ln -sf "$EMACS_APP" /Applications/Emacs.app
 
   '';
